requirements.txt
Flask>=3.1.1
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
psutil==7.0.0
flask_apscheduler==1.13.1
mcstatus==12.0.2
SQLAlchemy>=2.0.42
Flask-APScheduler>=1.13.1
python-dotenv==1.0.1